requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
Pillow==10.0.1
python-decouple==3.8
django-cors-headers==4.3.1
yookassa==2.4.0
requests==2.31.0
drf-yasg==1.21.7
django-filter==23.3
